test: split import meta glob tests
